Задача: Поисковый мини-движок Google. Инкапсуляция логики в классе WebPageService.
